Big refactor and dependency update.
Set up ESLint to have a more consistent coding style, and fixed the majority of issues with the code.

Also updated all the dependencies to the newest version. Removed git-user-name in the process, because it was using the
path library, whereas this is explicitly excluded from react 5. Did not see the need to have have git-user-name be the
default for an assignee.