Move state management into Board from App, and start rendering Board directly.
Also discovered an issue where the KanbnBoardPanel would start sending messages to its webview before the react
component (and more specifically the message listener) had been constructed. The solution was to request an update from
the component once everything is set up. This might end up being a problem with the other webviews as well, so will
likely use these solutions in that case as well.